Compute meanAndVar in constant space.
This is a typical laziness pitfall; here, the foldl' computing
the accumulator strictly means only that the tuple constructor is
evaluated, leaving the contents of the tuple as thunks, which then
nest as deep as the list.  The new version is strict in each of the
accumulators in the tuple.